Utility and KeyboardButton Modules
Moved _load_language, _xlate, and _strip_entities functions to a utility module. Language is loaded into the utility module to curb dependency on the Searcharr class for translation function in commands or other modules.

The KeyboardInlineButton definitions were moved to a new buttons module to make the _prepare_reponse and _prepare_user_response functions more readable.

Revised the calculations for _prepare_response_users pagination to eliminate offset and num so that it can use the same navigation button code as _prepare_response.